Pre-work for better isImplicate : removal of surfFunctor

+++ System/monteInc/Algebra.h
@@ -95,8 +95,7 @@ class Algebra
   std::ostream& write(std::ostream&) const;
   std::string writeMCNPX() const;
 
-  // Debug Functions::
-  int countLiterals() const;
+  size_t countLiterals() const;
 
 };
